requirements.txt
python-docx>=0.8.11
mistune>=3.0.0
Pillow>=10.0.0
pytest>=7.0.0
